pyproject.toml
[tool.poetry]
name = "data-emulator"
version = "0.1.0"
description = ""
authors = ["siakon89 <siaterliskonsta@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
SQLAlchemy = "^1.4.43"
psycopg-binary = "^3.1.4"
Faker = "^15.2.0"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"

[tool.poetry.dev-dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
